Include recent searches option in screen reader (#43)
Update screen reader text in search bar to include number of recent search options, and announce the option type when navigating through the option (e.g. a recent search or an autocomplete option)

J=SLAP-1792
TEST=manual

turn on screen reader on Mac, see that the screen reader announced the right number and type of option for recent searches and autocomplete option.