Add Dropdown compound component (#45)
This component is based on the dropdown components from [react-bootstrap](https://react-bootstrap.github.io/components/dropdowns/) and [restart/ui](https://react-restart.github.io/ui/Dropdown), which are uncoincidentally maintained by mostly the same people.

The useRootClose and useGlobalListener hooks are taken from directly from restart/ui.
- useRootClose gives modal like behavior to the dropdown, so that the dropdown closes when a click
is done outside of it, and also when the escape key is pressed.
- useGlobalListener adds a global listener to the document.

We could not copy react-bootstrap's approach for maintaining focus on a particular DropdownItem, which under the hood uses restart/ui's base Dropdown component.
This is because restart/ui's base Dropdown maintains focus by using the browser's native focus functionality.
This works great when you don't have an input, but with an input we want the browser's focus to be on that input 100% of the time, and the arrow keys to navigate a sort of "pseudo input"

This was remedied using a `numItems` prop, representing the total number of DropdownItems to be rendered, and also assigning each instance of DropdownItem an `index` prop. An alternative solution would be to decorate props.children like InputDropdown currently does.

J=SLAP-1777
TEST=manual

refactored the SearchBar to use the new Dropdown components (not part of this PR)
can navigate items
can loop through items using the arrow keys (both up and down)
autocomplete works
near me searches work
can run searches
did not notice any styling regressions (aimed to make DOM structure match the current InputDropdown)

did not test visual autocomplete

screenreader reads the annoucement text when the searchbar is clicked
used this [screenreader](https://chrome.google.com/webstore/detail/screen-reader/kgejglhpjiefppelpmljglcjbhoiplfn) (thanks @nmanu1 )
screenreader reads the SCREENREADER_INSTRUCTIONS only when clicked on and not when navigating dropdown items

----------

more notes on restart/ui's dropdown:
In restart/ui,
[First](https://github.com/react-restart/ui/blob/adf9d2d4835756b62e842f5567eb4be744291dce/src/Dropdown.tsx#L297), a native `querySelectorAll` is ran for all child elements with a specific magic data- attribute, and the component gets the element in the NodeList with an index of one more (or less, depending on which arrow key is pressed) than the element that currently has focus.
[Then](https://github.com/react-restart/ui/blob/adf9d2d4835756b62e842f5567eb4be744291dce/src/Dropdown.tsx#L298), the element is manually focused.

This does not work with a SearchBar because we want the browser's focus to be maintained on the `input` element 100% of the time. Otherwise, pressing an arrow key would result in leaving the input and not being able to type anymore.
We need to be able to maintain a "pseudo-focus" state like we currently do with InputDropdown.