Update screenreader announcements (#65)
Update ScreenReader announcements to include entity previews now that they are navigable by arrow key. Fix a bug where entity previews weren't displayed if there were no recent searches or query suggestions.

Also noticed a bug with the announcements while testing and fixed it here by using a separate variable to trigger the announcement instead of `value`. When moving focus from a dropdown item back to the input when the text for both is the same, the announcement was not made because the `value` in the input bar remained the same and `value` was being used as the `announcementKey`. For example, typing "cawfee", arrowing down to the autocomplete suggestion, and then back up to the input bar would not give any screen reader announcement.

J=SLAP-1858
TEST=manual

Check that entity previews are counted in the announcement when present.