Cut over UniversalResults (#15)
Cut over the UniversalResults component and add docs. Replaced `Link` with `<a>` tag so "View All " doesn't depend on react-router. Commented out the code for results count since currently the functionality for this is not being added to the component library.

J=SLAP-1876
TEST=manual

Link the component library to the starter app. Use UniversalResults and check that it looks as expected and the View All link works.