Scope package and update packages (#16)
Scope the package name so that we can publish to the Yext org on NPM

Also move react-router-dom from a devDependency to a dependency. Note, react-router-dom is different than react-dom which should be a peer dependency. With react-router-dom as a devDependency, it wasn't getting installed when the pagebuilder prototype was importing the search bar which was causing a runtime error. When we change it to a regular dependency, react-router-dom will get installed when a third party library specifies this package as a dependency.

J=none
TEST=manual

Publish to NPM and import the search bar in the PB prototype