Update Card Interface (#18)
Simplify the card interface by removing additionalProps

The only known prop of a card is now the Result which is passed to it. Any configuration of the card can be done with a HOC. This allows us to avoid passing around the card props, and it allows us to not have to type check them with unknown. Here's an example HOC with custom config (custom field mappings):

```tsx
({ result }: CardProps) => {
    return <StandardCard 
        result={result}
        fieldMappings={{
            title: {
                mappingType: 'FIELD',
                apiName: 'name'
            },
            description: {
                mappingType: 'FIELD',
                apiName: 'venueName'
            }
        }}
    />
}
```

J=SLAP-1857
TEST=manual

I tested this code in the feature/prototype-answers-in-PB branch of the starter app. I tested that I was able to configure cards by creating a HOC which passes in custom props to the standard card.