Add API Extractor (#25)
Add API Extractor for automatic documentation generation

Use the same package.json updates as Answers Headless. Also null check the analytics lib in order for the lib to build. 

Additional changes in order to address all of the warnings that API Extractor logs will be addressed in the Documentation Review Item (SLAP-1897).

In the documentation review item, we should update the package.json file to point the the generated type bundle so that the exported types will respect the release tags such as `@beta` or `@internal`.

Note: The documentation validation will not pass until the dev/documentation-review item is complete

J=SLAP-1888
TEST=manual

Run a build and inspect the generated docs in the docs folder