Documentation review (#32)
Update documentation for components

For this PR, I went through the API report warnings and errors and addressed them. This involved adding missing documentation for component props and updating existing docs to work with API extractor. In several spots I needed to change the `inheritDoc` tag to `link` because they were wiping out the entire doc summary.

Filters is exported as a namespace which works a little differently than our other components. API extractor gives warning about unsupported link types when linking to the namespaced Filters members, however the links in the generated docs work just fine, so I opted to disable the warnings in the CLI and instead move them to the API report file so that we don't totally ignore the missing links, but also so that the production API extractor build can exit with an exit code of 0.

Currently the docs for models which come from @yext/answers-headless-react and @yext/analytics are not working. This is expected for @yext/analytics because we don't specify a type file for that repo. I attempted to bundle the packages just like we do in answers-headless, however the docs weren't being generated. I believe we may need to update how our other libs export types in order to get it working. Therefore, deferring to a new item.

API documenter does a better job of documenting interfaces than types, so I switched a few of them.

There is a conflict for `VerticalResults` since it is both a data model and a component. This will be resolved in SLAP-1899

J=SLAP-1897
TEST=visual

Go through the generated docs and confirm they look correct visually. Ensure I can navigate through the links as expected.