Vertical config interface update (#34)
Consolidate our vertical config interfaces around a `VerticalConfigMap` interface.

Previously we had a verticalsConfig object and a verticalConfigs object where one is an array and the other is an object keyed by vertical key. This consolidates on a single object so that we can pass the same config to both the UniversalResults and the AlternativeVerticals component. The alternative verticals component relies on the `VerticalLabelMap` which is a subset of the `VerticalConfigMap`. I'm thinking we should use the same prop name `verticalConfigMap` to help indicate that the same data structure can be passed to both components.

J=SLAP-1899
TEST=none

Import this component in the sample app and smoke test the alternative verticals component. Ensure that it looks the same as the site on develop.