Migrate from default exports to named exports (#36)
Remove all default exports and make them named exports

The primary reason for using named exports is that the component's file is the source of truth for its own naming. Since we are re-exporting many models and interfaces, the actual name of the component was defined in the index.js files where it was being re-exported. By switching to named exports, the source of truth can live in the component's file itself, and we reduce code duplication.

I also removed `"isolatedModules": true` which lets us remove the `type` from our exports. The `type` keyword is primarily used by compilers other than typescript itself (such as Babel) which compile typescript one file at a time and don't have the type information to know whether or not an export is a type. The Typescript compiler has access to this information, so therefore we can disable "isolatedModules".

J=SLAP-1901
TEST=manual

Ensure typescript builds the library successfully. Smoke test with the sample app using the Filters component.