Reset pagination when filters are applied (#42)
For Facets, reset pagination when applying a filter when searchOnChange=true. For StaticFilters, reset pagination when a filter is clicked. Also reset pagination when an AppliedFilter is removed.

The apply button support is still in progress, so currently this PR doesn't reset pagination when searchOnChange=false for Facets.

J=SLAP-1903
TEST=manual

Connect these components to the react site search starter and set the verticalLimit to a low number such as 1 or 3. Navigate to a page past 1 and apply a filter. See that the pagination is reset when a filter is selected. Set `searchOnChange=false` for Facets and see that pagination is not reset in that case. Test that removing an applied filter resets pagination as well.