Switch from blue to primary color (#43)
Update the usage of the blue color to the custom primary color.

This means that any application using these components must define the primary color in the tailwind config. For example, to use the default blue, the tailwind.config.js should include:

```js
theme: {
    extend: {
        colors: {
            primary: colors.blue,
        }
    }
},
```

The `colors` object must be imported at the top of the config with:
```js
const colors = require('tailwindcss/colors')
```

J=SLAP-1905
TEST=manual

Connect the filter component to the sample app and set the primary color in the config. See the color update on the web page with hot module replacement.