Add Jest tests for the `SpellCheck` Component.
This PR adds the first of the Component Jests tests to the library.
The following scenarios are tested for `SpellCheck`:

- The suggestion text is formatted properly.
- The button label is correct.
- The `onClick` handler is invoked, if provided.
- The `executeSearch` method is invoked if no `onClick` is present.

As part of this PR, I did a good amount of investigation into how to
mock a custom hook. For now, I haev the `answers-headless-react` hooks
mocked in the test file itself. If we find that we are re-using the same
mock implementations in other tests, we can re-factor the mock into a
stand-alone fixture. There is a potential alternative to my approach with
these mocks. Specifically, we could mock the `AnswersHeadlessContext`
itself and make a Provider for the mocked Context. I believe the approach
I have now is simpler, however.

I also figured out a way in which we can spy on invocations of the mocked
Answers Actions. It is a little complicated because we are not injecting
the hooks as props.

J=SLAP-1921
TEST=auto