Add Applied Filters Tests (#53)
Add test coverage for Applied Filters. Test that the filters render filters in headless state, and test that the clear buttons work. Add aria labels to the remove buttons.

When mocking @yext/answers-headless-react, I'm including the original module so that I can access various enums and interfaces including the Matcher. 

Adding "transformIgnorePatterns" allows '@yext/answers-headless-react' to be transformed and used in the jest test. Otherwise it throws an error because the library uses ESM rather than CJS.

J=SLAP-1922
TEST=auto

Run tests and confirm the pass. See that the code coverage numbers are between 90 and 100%