Add Storybook and first Story for the `SpellCheck` Component.
This PR adds Storybook to the Component library. Storybook, in
conjunction with Percy, will provide Component-level visual regression
tests.

In this PR, I've added a single Story for the `SpellCheck` Component.
A subsequent PR will handle the full Percy integration. For now, the
snapshots generated by Storybook are not uploaded to our Percy project.

J=SLAP-1928
TEST=manual