Add GH Workflow for Percy Snapshots
This workflow is run on all pull requests and most pushes. It first generates the Storybook build, which contains the Stories for each Component. Percy then takes this build and generates Snapshots which are uploaded to the `answers-react-components` Project.

J=SLAP-1928
TEST=manual

I tested the steps of the `default-snapshots` job manually and verified that this resulted in the expected Snapshots being uploaded to the Percy Project.