Add Percy Tests for `SpellCheck`.
This PR adds Storybook to the Component library. Storybook, in conjunction with Percy, will provide Component-level visual regression tests. I've added a single Story for the `SpellCheck` Component. As part of creating the Story, I added a number of test fixtures and mock generators, which allow us to construct a controlled instance of Headless. 

A `percy-snapshots` GH workflow was added. This workflow is run on all pull requests and most pushes. It first generates the Storybook build, which contains the Stories for each Component. Percy then takes this build and generates Snapshots which are uploaded to the `answers-react-components` Project.

J=SLAP-1928
TEST=manual