Address most eslint-plugin-react-perf errors (#61)
Address 51 eslint errors, the majority of which are from eslint-plugin-react-perf

These changes improve the performance of the page by preventing unnecessary re-renders. The simplest fix is generally to move the object or function out of the component itself. Sometimes that is not possible because the callback depends on  variables defined within the component. In that case, the idiomatic solution is to use `useCallback` or `useMemo`. The case where new callbacks are being created within a map/loop are more difficult to address because `useCallback` must be used at the top level of a React component. I fixed one of these cases by creating a new react component which is rendered within the loop which calls the `useCallback` hook within the component. The remaining 5 remaining linting errors involving loops will be addressed in a future item. We should also create another item for optimizing our hooks. I don't believe the plugin properly detects new objects/functions being created within hooks that the callbacks may depend on.

I also created `useCardAnalyticsCallback()` and `useCardFeedbackCallback()` as convince functions for creating memoized analytics callbacks.

J=SLAP-1940
TEST=none

Use a local version of the pagebuilder prototype with these components and smoke test basic functionality