VA Styling Tweaks (#63)
Adjust the VA text color for autocomplete suggestions and for recent searches. Also fix the dropdown shadow.

Set the text color to 'text-gray-800' as indicated by product.

I also noticed that the search bar dropdown shadow wasn't covering filter search on a vertical page. I fixed this by moving the 'relative' position to the inputDropdownContainer and updating the z-index.

The item also called for the icons to the left of the query to be the same color. This was fixed a while back so no updates are necessary here.

J=SLAP-1959
TEST=manual

View the DOM in devtools and confirm the color is 'text-gray-800' for both recent searches and regular autocomplete suggestions. See that the search bar shadow properly covers the filter search component and the edge of the nav bar.