add result count component for universal and vertical (#67)
Add support to display results count for universal and vertical search. The mocks have results count display outside of the VerticalResults component. Instead of embedding the count in UniversalResults/VerticalResults, I have it as its own component that can render outside of the results component.

Small css change for UniversalResults to follow the styling pattern we have in most of the components where we intentionally use margin-bottom so it's consistent.

J=SLAP-1886
TEST=manual

test in theme 3.0 prototype, and see that results count appears in universal and vertical page as expected.