use snapshotted filters state for AppliedFilters
Updates AppliedFilters to use the filters state snapshotted
from after the last search ended.

Previously, AppliedFilters would update the filter state whenever
a search was not loading. This could result in filters showing up as
applied before they were applied to the current search. This can be
seen by setting searchOnChange to false for Facets or StaticFilters.

J=SLAP-1984
TEST=manual

clicking a Facet with searchOnChange=false does not add it to the
AppliedFilters until a search has been run with it