Hierarchical Facets QA (#72)
Rework Hierarchical Facets so that only one facet is selected at a time

Previously Hierarchical facets relied on all parents of a facet to be selected. This updated updates the rendering logic,  callbacks, applied filters, and tests to account for this updated behavior

J=SLAP-1975
TEST=manual, auto

Manually test the various workflows of clicking available options, current categories, parent categories, and the all categories button. Add test to ensure the correct behavior when clicking on the various buttons of the component