Tweak VA hover state styling (#79)
The styling for the hover state was already present. This PR makes a small change to the styling of the `SearchBar` component. By making the top and bottom padding of the `dropdownContainer` consistent, when there is only one dropdown option, the hover background looks centered. Entity preview styling is controlled by users of the `SearchBar` component, such as `SampleVisualSearchBar` in the starter app, so no changes are needed in the component lib to support that.

J=SLAP-1982
TEST=manual

Check the Theme 3.0 test site locally and see that the hover state looks as expected for recent searches and autocomplete suggestions.