support highlighted fields in StandardCard's description
Added a new applyHighlightedFieldDataPath, which operates
similarly to applyFieldDataPath, but has additional logic
to first check the results highlightedFields.

If no highlightedField is found, it will default to the rawData.
The field value found in the rawData will then be formatted into
a HighlightedValue object with 0 matchedSubstrings.
Originally, I had it just return a string in this case, but this made
the data validation as well as render logic in the StandardCard more
complicated.

J=SLAP-1988
TEST=manual

I can see highlighted fields in the test-site's cards' descriptions
search for "kazuma" and "lorem" on the people page
descriptions without highlighted fields also display properly