Fix react-perf looping errors (#85)
Fixes most of the errors found from `eslint-plugin-react-perf` that were caused by looping. The only remaining errors are caused by how `itemData` is passed to `DropdownItem` components. This will require a refactor of the `Dropdown` components to remove the `itemData` concept and will be part of a different item.

J=SLAP-1964
TEST=manual

Hook up locally to the Theme 3.0 test site and make sure applied filters work as expected.