Add `Primary` Story for `AppliedFiltersDisplay`. (#90)
This PR adds the Storybook visual regression test for the `AppliedFiltersDisplay` Component.
Note that the display Component was used in place of `AppliedFilters`. I feel all of our
Storybook tests should be written against the various Display Components, as opposed to their
parents. Our Jest tests should ensure that the parent Componetns are functioning properly. This
includes verifying that the child Display Components are invoked with the right props. Therefore, all
that's left to test in Storybook is that the Display looks correct given the props.

J=SLAP-1991
TEST=auto

Ensured all unit tests and the Storybook build passed. Saw the new Story I added appear as expected.