highlighted fields in StandardCard, remove applyFieldMappings (#86)
This commit adds highlighting to the StandardCard's title and description sections.

It also removes applyFieldMappings from the component library.
applyFieldMappings was removed because it was arguably an implementation detail of the
answers-react-theme. More importantly, though, it wasn't the most clear interface for a user
using the component library in a standalone fashion. 
Instead, we now just have a function that does this, called dataForRender.

J=SLAP-1988
TEST=manual

I can see highlighted fields in the test-site's cards' descriptions
search for "kazuma" and "lorem" on the people and universal test-site pages
descriptions without highlighted fields also display properly

checked that an undefined description is handled properly with the code snippet

```ts
  const transformedFieldData = applyFieldMappings({
    ...result.rawData,
    description: undefined
  }, {}, {
    ...defaultFieldMappings,
    ...customFieldMappings
  });
```

checked with Rose on the styling