reset offset on filtersearch option select (#95)
reset offset for a fresh search when user select an option on filtersearch.

J=SLAP-2019
TEST=manual

update test-site People page to use FilterSearch. Perform an empty search to receive all the results, click on second results page on pagination, use filterSearch to select Tom M. See that:
without the change: no result is displayed even though results count is 1.
with the change: the expected one result is displayed on page.