Add RangeInput (#92)
Add RangeInput

Add the RangeInput which is a static filter which can be rendered with numeric facets. RangeInput by default works with regular numbers, however it can easily support currencies. In order to use it with USD, a new getFilterDisplayName function must be supplied which uses a dollar sign. Also, the inputPrefix must be supplied with a JSX element containing a dollar sign.

Adding the tooltip will be part of a future PR.

TEST=manual

Add a new page to the test site and test that numbers can be entered, and that they affect the state in headless and affect the filters sent in the request. Test unbounded filters as well and confirm they show up in the applied filters. Test that the 'clear min and max' button clears the inputs. Test that the 'apply' button clears selected static filters with the same fieldId, selects the current filter, and fires off a search