Add WCAG Accessibility testing to each Story.
This PR adds the `a11y` addon to our Storybook infrastructure. This
addon allows us to run WCAG tests against the DOM of each of our stories.
Currently, the addon is configured to test against the same WCAG standards
that the Theme does.

J=SLAP-2031
TEST=manual

Verified that in the Storybook build/preview there were accessibility reports
for each Story.