Add WCAG Accessibility testing to each Story. (#99)
This PR adds the `a11y` add-on to our Storybook infrastructure. This add-on allows us to run WCAG tests against the DOM of each of our stories. Currently, the add-on is configured to test against the same WCAG standards that the Theme does. In another PR, I will figure out how to report the WCAG results directly on PRs. 

J=SLAP-2031
TEST=manual

Verified that in the Storybook build/preview there were accessibility reports for each Story.