Add invalid range messaging (#98)
Add a message for invalid ranges

If the min is greater than the max, show the invalid message. When the input is invalid, the apply button should do nothing. I also discussed the situation where one or both of the inputs is supplied with a '.'. In that case, the invalid message should not show, but the apply button should not try to send the facet to the backend.

J=SLAP-2025
TEST=manual

Type a larger number in the min than in the max and see the invalid message appear. See that when a '.' is entered, the invalid message doesn't show, but the apply button doesn't try to apply the filter.