Make answers-headless-react a peer dependency (#100)
The purpose of making it a peer dependency is that it allows the components to work with the answers-headless-react version of the application. 

If there are two answers-headless-react libraries, then the react hooks will cause a runtime error. Note, answers-headless-react is also a devDependency, so it will be installed for development.

When building an app which depends on this library, answers-headless-react will be automatically installed for NPM versions 7 or newer. If someone is on NPM 6, then will need to add answers-headless-react as a dependency of their application.

J=SLAP-2010
TEST=manual

Test with the test site that the app works as intended. It's difficult to test the peer dependency since the module resolution works differently when the package is actually published to NPM, however I'm pretty confident it will work as intended based on my understanding of devDependencies and peerDependencies. I can publish a beta after this is merged to confirm this works as intended