Add Jest tests for Facets (#102)
Adds Jest tests for the `Facets` component for both regular and number range facets. The tests check the selecting and unselecting behavior of facet options and the behavior of `searchOnChange`. Only number range facet checkbox options were tested. Testing for the range input and disabling interactions are not part of this item.

J=SLAP-2032
TEST=auto

See that the newly added tests pass.