Perform a single CSS composition (#119)
This PR ensures that components are only performing a single CSS composition. A child component either relies solely on the parent components to perform the composition and now accepts only `cssClasses`, or if it might be used without a parent, it performs it's own composition, but parent components pass in 'replace' for the `cssCompositionMethod`.

J=SLAP-2011
TEST=manual

Spin up the test-site and click through to make sure the styling is still applied as expected.