SSR Support for Yext Sites Dev Build (#125)
Add SSR support to the YSS dev build by adding `"type": "module"` and useIsomorphicLayoutEffect

By setting `"type": "module"`, node treats the javascript files within the package as ES modules. Since some of the config files are CJS, we need to indicate that by updating the file extension from '.js' to '.cjs' or else node 16+ will throw an error

J=none
TEST=manual

Connect to the Yext Sites Starter and smoke test. Note: There appears to be an issue with Yext Sites when this code is downloaded from NPM (or the bundle from npm pack is used), however that issue appears to be on the Sumo side