Refactor the `StandardCard` and add `PropTypes`. (#124)
This PR splits the `StandardCard` into two Components: `StandardCard` and `StandardCardDisplay`. This was done to enhance readability and organization. There was a lot going on in the original `StandardCard`, making the file hard
to parse. We want readability so that users of the library easily know how to make their own cards. In the `StandardCardDisplay`, I also removed a number of the CSS configuration options. The amount of CSS customization we had was unnecessary. People are unlikely to use this Card, instead it will serve as an example of how to create a Card. To get our point across, we don't need every element of the DOM to have configurable styling. 

I removed our own `validateData` run-time type checking from the card. Instead, I used the native `React.PropTypes` in the display Component. This will be more natural for React developers. `PropTypes` is the idiomatic way to do type
checking at run-time in React.

TEST=auto

Ensured all the unit tests passed.