Add Storybook controls (#153)
This PR adds Storybook controls for each component's stories. This allows almost all component props to be configured from the Storybook UI. Props that are functions or objects containing functions (e.g. `onSearch`, `onClick`, `renderEntityPreviews`, etc.) cannot be directly configured via the controls (see https://storybook.js.org/docs/react/essentials/controls#dealing-with-complex-values).

Additionally, there are a few props that can be set, but their values don't affect the components because we are hardcoding the returned response results. I disabled the controls for these, but if we want to update the response based on those props in the future, they can easily be added back:
- `hierarchicalFacetsDelimiter` in `AppliedFiltersDisplay`
- `sectioned` and `searchFields` in `FilterSearch`
- `geolocationOptions` in `LocationBias` and `SearchBar`

The `expanded: true` config option in `.storybook/preview.js` ensures that the prop descriptions remain visible in the table on the Storybook site even after controls are enabled.

Lastly, I noticed that the applied filters styling was not being applied for universal results, so the `builtInCssClasses` in `UniversalResults` was updated to fix that and the corresponding stories were updated to display the applied filters by default.

J=SLAP-2094
TEST=manual

Went through each component's stories and checked that each prop is configurable using the controls and changes the display as expected except for those exceptions listed above.