Add `StaticFilters` component with baked-in layout (#156)
This PR adds a higher-level `StaticFilters` component that uses the existing compound components and has a baked-in layout. Only basic static filters (with an `$eq` matcher) are supported. Each filter option in this component will have the same `fieldId`, so a page might have multiple `StaticFilters` components to filter on multiple `fieldId`s.

While testing, I found a bug with our current implementation of selecting a filter by default. The Clear All button of applied filters would clear out the static filters entirely, so each time the static filters that were selected by default would be applied again even after clearing them. This was fixed by updating the Clear All button to unselect static filters instead of clearing them out of state.

The `CollapsibleSection` styling was updated to ensure the apply filters button styling remained as is even when included in a `CollapsibleSection`. Removing the flex styling didn't have any other effects on desktop or mobile.

Lastly, since we eventually plan to not export any of the `Filters` compound components, I commented out the `StaticFilters` import in `Filters/index.ts` but left it in the file. Once we can remove all exports, I think it would be easier to remove the `Filters` export from `components/index.ts` and uncomment the exports in `Filters/index.ts` once more. This way, internally we'll be able to import all of the compound components from the same `Filters` module path even if we no longer export them for users.

J=SLAP-2105
TEST=auto, manual

Check that Jest tests pass and test each prop for `StaticFilters` with the test-site to ensure correct behavior.