Update Storybook site sync workflow (#165)
Updates the `sync-sites-branch` workflow to automatically merge `main` into the `storybook-site` branch whenever there is a push to `main`. This uses a new GitHub action: [mtanzi/action-automerge](https://github.com/mtanzi/action-automerge). Now, instead of making a PR that has to be manually merged in, the bot will make a merge commit automatically. If there is a merge conflict, the workflow will fail, in which case the conflict would need to be resolved manually and merged into the `storybook-site`.

I had originally looked into whether the site config could be kept on `main` so we wouldn't need to maintain a separate `storybook-site` branch, but it doesn't seem like there is a way to avoid having those config files included when someone outside of the team clones the repo.

J=SLAP-2119
TEST=manual

Fork the repo, update the workflow, and make commits to main. See that they are automatically merged into the storybook-site branch. Test that the workflow fails if the merge is not successful (e.g. merge conflict).