VisualAutocomplete: delete EntityPreviews (#168)
This PR deletes the EntityPreviews component. Originally, the purpose was to only remove automatic detection of universalLimit and restrictVerticals based on props passed into EntityPreviews. However, there really was not a lot of
purpose for EntityPreviews to exist after doing so, so it was removed.

I intentionally made restrictVerticals a required prop. People should not be using visual autocomplete
without setting it - we're already bombing the universal search endpoint with this feature.

universalLimit was kept as optional.

J=SLAP-2118
TEST=manual

add VA to the test-site
see it display properly

tested that the screenreader correctly displays the entity previews count by inspecting it through react dev tools