Make `builtInCssClasses` readonly (#176)
This PR makes each component's `builtInCssClasses` readonly.

I opted to use the Typescript `Readonly` utility type which adds `readonly` to each of the properties of an object. Although, this is a shallow operation, in the case of the `builtInCssClasses`, this is sufficient because the object will only contain properties with string values. I initially tried using const assertions, but because we define the type for the classes, such as `const builtInCssClasses: SearchBarCssClasses = { ... }`, it seems to override the const assertion and allows the properties to be modified.

J=SLAP-2136
TEST=manual

Tried modifying the classes in `builtInCssClasses` and saw it gives an error about not being able to assign to read-only properties.