Add placeholder prop to FilterSearch (#177)
Add `placeholder` to `FilterSearchProps` to allow the search input's placeholder text to be customized. Update the default value to match `SearchInput` and update the name of the prop in `SearchInput` from `placeholderText` to `placeholder` to match the `SearchBar`.

J=SLAP-2141
TEST=auto, manual

Update the Jest tests to verify the default value of the placeholder text and to make sure the text is customizable. Spin up the test-site and see that the placeholder text appears and can be customized as expected.