Remove old filter values (#185)
This changes FIlterSearch such that you can only have one filter at a time. When you add the first filter, it is set as a filter option for the next search. Upon further addition of searched filters, the old ones will be replaced by the newly searched ones. Note that backspacing from the filter box does not remove the filter (just as ZocDoc does for its doctor filter).

J=SLAP-2142
TEST=auto

One jest test was added to test the added functionality.