Add GH Action to run ES Linter on PRs into `main`. (#199)
Set `--maxWarnings=0`, so ESLint will treat all warnings as errors for extra safety. I addressed the three existing warnings already in the repository. In the future, we will look to make this action part of slapshot-reusable-workflows.

TEST=manual

Set ESLint Warnings to Errors and saw the PR check fail. Reverted this change and saw the PR check pass.