added hover styling to option, similar to SearchBar (#203)
This PR makes hovering over filter search autocomplete suggestions trigger a highlight on the selected option.  The JIRA item requested that the `focusedOption` CSS should be applied to options that are hovered over.  However, adding a `hover` style to `option` more closely replicates what is done in `SearchBar` and seems to be a more straightforward approach.

J=SLAP-2179
TEST=manual

Attempted to add hovering test to storybook, but ran into issues with simulating hover.