Add Yext branding to the Storybook site (#211)
This PR makes some changes to the branding of the Storybook site to match some of the colors and styles from our components using Storybook theming. The title of the site has been updated and includes the Yext logo, and clicking on the title opens a new tab with the Github repo. The Yext logo is also displayed on the browser tab.

J=SLAP-2176
TEST=none

<img width="1295" alt="Screen Shot 2022-06-29 at 12 33 11 PM" src="https://user-images.githubusercontent.com/88398086/176489016-f6977733-0397-4243-8ded-8bfc78feb729.png">