fixed dropdown menu tabbing (#207)
This PR addresses tabbing issues present in dropdown menus.  Now, when the user hits `tab` when a dropdown is present, the next dropdown element will be selected.  `shift+tab` will do the opposite.  Upon exceeding the last element in the dropdown menu, the dropdown menu will be closed and the next component after the dropdown will be focused.

SLAP=J-2193
TEST=auto, manual

A Storybook story was added for FilterSearch and SearchBar to test highlighting, and two unit tests were added to dropdown to test that tabbing can be used just as arrow keys, and that you can tab out of a dropdown menu which causes it to close.